Move classes to rxbinging2 package
