+++ rxbinding-recyclerview-v7/src/androidTest/java/com/jakewharton/rxbinding2/support/v7/widget/RxRecyclerViewAdapterTest.java
@@ -1,4 +1,4 @@
-package com.jakewharton.rxbinding.support.v7.widget;
+package com.jakewharton.rxbinding2.support.v7.widget;
 
 import android.support.test.annotation.UiThreadTest;
 import android.support.test.rule.UiThreadTestRule;
