+++ rxbinding-recyclerview-v7/src/androidTest/java/com/jakewharton/rxbinding2/support/v7/widget/RxRecyclerViewTest.java
@@ -1,4 +1,4 @@
-package com.jakewharton.rxbinding.support.v7.widget;
+package com.jakewharton.rxbinding2.support.v7.widget;
 
 import android.app.Instrumentation;
 import android.support.test.InstrumentationRegistry;
